+++ src/history.js
@@ -1,0 +1,17 @@
+let history = [];
+
+export const add = (value) => {
+  if (!value || history.includes(value)) {
+    return;
+  }
+
+  history.push(value);
+};
+
+export const pop = () => {
+  return history.pop();
+};
+
+export const reset = () => {
+  history = [];
+};
