+++ src/index.js
@@ -1,130 +1,14 @@
-const install = (Vue, newProps = {}) => {
-  const history = [];
-  let notificationsCount = 0;
-  const $page = { title: "" };
-  const defaultProps = {
-    divider: "-",
-    title: null,
-    router: null,
-  };
-  const props = { ...defaultProps, ...newProps };
+import { updateOptions } from "@/options";
+import { defineTitleProperty } from "@/title";
+import mixin from "@/mixin";
+import addRoutesTracking from "@/add-routes-tracking";
 
-  const { router } = props;
+const install = (Vue, options = {}) => {
+  updateOptions(options);
+  addRoutesTracking();
+  defineTitleProperty(Vue);
 
-  const setPreviousTitle = () => {
-    history.pop();
-    setTitle(history.pop());
-  };
-
-  const setPageTitle = (title) => {
-    const { title: appName, divider } = props;
-
-    let pageTitle = document.title;
-
-    if (title) {
-      pageTitle = title + (appName ? ` ${divider} ${appName}` : "");
-    } else if (appName) {
-      pageTitle = appName;
-    }
-
-    if (notificationsCount > 0) {
-      pageTitle = `(${notificationsCount}) ${pageTitle}`;
-    }
-
-    document.title = pageTitle;
-  };
-
-  const addToHistory = (title) => {
-    if (!title || history.includes(title)) {
-      return;
-    }
-
-    history.push(title);
-  };
-
-  const setTitle = (title) => {
-    setPageTitle(title);
-    addToHistory(title);
-
-    $page.title = title;
-  };
-
-  Object.defineProperty(Vue.prototype, "$title", {
-    get: () => $page.title,
-  });
-
-  Vue.mixin({
-    data() {
-      return {
-        $_vuePageTitle_isTitleSet: false,
-      };
-    },
-
-    created() {
-      const { title } = this.$options;
-
-      if (!title) {
-        return;
-      }
-
-      this.$setPageTitle(
-        typeof title === "function" ? title.call(this, this) : title
-      );
-    },
-
-    beforeDestroy() {
-      this.$resetPageTitle();
-    },
-
-    methods: {
-      $updateNotificationsCounter(count) {
-        notificationsCount = count;
-        setTitle(history.pop());
-      },
-
-      $resetPageTitle() {
-        if (this.$_vuePageTitle_isTitleSet) {
-          this.$_vuePageTitle_isTitleSet = false;
-          setPreviousTitle();
-        }
-      },
-
-      $setPageTitle(title) {
-        this.$_vuePageTitle_isTitleSet = true;
-        setTitle(title);
-      },
-    },
-  });
-
-  if (!router) {
-    return;
-  }
-
-  const getRouteTitle = (route) => {
-    const nearestRoute = route.matched.find((route) => route.meta.title);
-
-    let title;
-
-    if (route.meta.title) {
-      title = route.meta.title;
-    } else if (nearestRoute) {
-      title = nearestRoute.meta.title;
-    }
-
-    return title;
-  };
-
-  router.onReady(() => {
-    setTitle(getRouteTitle(router.currentRoute));
-
-    router.afterEach((to) => {
-      if (to.matched.find((route) => route.meta.inheritPageTitle)) {
-        return;
-      }
-
-      setTitle(getRouteTitle(to));
-    });
-  });
+  Vue.mixin(mixin);
 };
 
 export { install };
